Reorganize tests/ directory
This commit cleans up some stuff that's always bothered me.

* Move runtime engine scripts into tests/runtime/engine
  * There was already a utils.cpp and parser.cpp that do very different
    things than utils.py and parser.py

* Move tests/btf_data.h to tests/data/btf_data.h to namespace test
  scripts and data

* Update tests/README.md with more up to date info
