+++ tests/CMakeLists.txt
@@ -181,15 +181,18 @@ endforeach()
 add_custom_target(testlibs DEPENDS ${compiled_testlibs})
 
 configure_file(runtime-tests.sh runtime-tests.sh COPYONLY)
-configure_file(main.py main.py COPYONLY)
-configure_file(parser.py parser.py COPYONLY)
-configure_file(utils.py utils.py COPYONLY)
 file(GLOB runtime_tests runtime/*)
+list(REMOVE_ITEM runtime_tests ${CMAKE_CURRENT_SOURCE_DIR}/runtime/engine)
 list(REMOVE_ITEM runtime_tests ${CMAKE_CURRENT_SOURCE_DIR}/runtime/scripts)
 list(REMOVE_ITEM runtime_tests ${CMAKE_CURRENT_SOURCE_DIR}/runtime/outputs)
 file(MAKE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/runtime/)
 foreach(runtime_test ${runtime_tests})
   configure_file(${runtime_test} ${CMAKE_CURRENT_BINARY_DIR}/runtime/ COPYONLY)
+endforeach()
+file(GLOB runtime_engine_files runtime/engine/*)
+file(MAKE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/runtime/engine)
+foreach(runtime_engine_file ${runtime_engine_files})
+  configure_file(${runtime_engine_file} ${CMAKE_CURRENT_BINARY_DIR}/runtime/engine/ COPYONLY)
 endforeach()
 file(GLOB runtime_test_scripts runtime/scripts/*)
 file(MAKE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/runtime/scripts)
